+++ src/TabularColumn.php
@@ -18,6 +18,11 @@ use yii\base\Model;
 class TabularColumn extends BaseColumn
 {
     /**
+     * @var TabularInput
+     */
+    public $context;
+
+    /**
      * Returns element's name.
      *
      * @param int|null $index current row index
@@ -63,16 +68,10 @@ class TabularColumn extends BaseColumn
      */
     public function setModel($model)
     {
-        $currentModel = $this->getModel();
-
-        // If model is null and current model is not empty it means that widget renders a template
-        // In this case we have to unset all model attributes
-        if ($model === null && $currentModel !== null) {
-            foreach ($currentModel->attributes() as $attribute) {
-                $currentModel->$attribute = null;
-            }
-        } else {
-            parent::setModel($model);
+        if ($model === null) {
+            $model = \Yii::createObject(['class' => $this->context->modelClass]);
         }
+
+        parent::setModel($model);
     }
 }